# platform/contracts/openapi/v1/wego-api.yaml
openapi: 3.1.1
info:
  title: Wego API
  version: 1.0.0
  description: >-
    Version 1 of the product-neutral Wego HTTP contract. WEGO-001 adds the
    first authenticated operations (login, logout, current identity);
    WEGO-002 adds the first business-domain operations (diving offerings and
    staff-created bookings).
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Current client instance
...
paths:
...
  /api/v1/tours-operator/staff/sales-control:
    get:
      operationId: getToursOperatorSalesControl
      summary: Emergency sales control for managers
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current switch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorSalesControl"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
    put:
      operationId: updateToursOperatorSalesControl
      summary: Pause or resume new online bookings and payment checkouts
      description: >-
        Requires tours-operator.tour:manage. Webhooks, booking expiry and
        staff operations keep running while paused.
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateToursOperatorSalesControlRequest"
      responses:
        "200":
          description: Switch saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorSalesControl"
        "400":
          $ref: "#/components/responses/ToursOperatorValidationResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque session token issued by POST /api/v1/identity/login.
  responses:
...
    UnauthenticatedResponse:
      description: Missing, invalid, or expired session token.
      headers:
        WWW-Authenticate:
          description: Always `Bearer` — the authentication scheme this API accepts.
          schema:
            type: string
            enum:
              - Bearer
    ForbiddenResponse:
      description: Authenticated, but missing the permission this operation requires.
...
    ToursOperatorValidationResponse:
      description: Malformed or domain-invalid tours-operator request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
...
  schemas:
...
    ToursOperatorBookingMode:
      type: string
      enum: [ONLINE_PAYMENT, ENQUIRY_ONLY]
      description: Deployment capability. Enquiries are not bookings and never reserve inventory.
    ToursOperatorSalesControl:
      type: object
      additionalProperties: false
      required:
        - bookingsPaused
        - paymentsPaused
        - bookingMode
      properties:
        bookingsPaused:
          type: boolean
        paymentsPaused:
          type: boolean
        bookingMode:
          $ref: "#/components/schemas/ToursOperatorBookingMode"
        reason:
          type: [string, "null"]
          maxLength: 300
        updatedByUserId:
          type: [string, "null"]
          format: uuid
        updatedAt:
          type: [string, "null"]
          format: date-time
    UpdateToursOperatorSalesControlRequest:
      type: object
      additionalProperties: false
      required:
        - bookingsPaused
        - paymentsPaused
      properties:
        bookingsPaused:
          type: boolean
        paymentsPaused:
          type: boolean
        reason:
          type: [string, "null"]
          maxLength: 300
...
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - validation_failed
        message:
          type: string
          description: A safe, human-readable description of which rule failed.
...